built the find_existing_score() and acceptable() functions
built functions to find the existing score of a comment because if a reply already exists for a comment and the reply has a better score, then we need to update the row with that new reply. We also need to make sure that the comment we are using has acceptable data.